general: Rewrite Field to decorator
