+++ package.json
@@ -37,40 +37,40 @@
   "author": "oreqizer",
   "license": "MIT",
   "devDependencies": {
-    "@types/enzyme": "~2.8.1",
-    "@types/jest": "~20.0.2",
-    "@types/ramda": "0.24.0",
-    "@types/react-redux": "~4.4.45",
+    "@types/enzyme": "~2.8.4",
+    "@types/jest": "~20.0.5",
+    "@types/ramda": "0.24.5",
+    "@types/react-redux": "~4.4.47",
     "@types/redux": "~3.6.0",
     "babel-cli": "~6.24.1",
     "babel-jest": "~20.0.3",
-    "babel-loader": "~7.1.0",
-    "babel-plugin-ramda": "~1.2.0",
+    "babel-loader": "~7.1.1",
+    "babel-plugin-ramda": "~1.3.0",
     "babel-preset-es2015": "~6.24.1",
     "babel-preset-react": "~6.24.1",
     "babel-preset-stage-3": "~6.24.1",
-    "chalk": "~1.1.3",
+    "chalk": "~2.0.1",
     "cross-env": "~5.0.1",
     "enzyme": "~2.9.1",
     "gulp": "~3.9.1",
     "gulp-babel": "~6.1.2",
     "gulp-plumber": "~1.1.0",
-    "gulp-typescript": "~3.1.7",
+    "gulp-typescript": "~3.2.1",
     "gulp-util": "~3.0.8",
     "jest": "~20.0.4",
-    "lerna": "2.0.0-rc.5",
+    "lerna": "2.0.0",
     "ramda": "~0.24.1",
     "react": "~15.6.1",
     "react-addons-test-utils": "~15.6.0",
     "react-dom": "~15.6.1",
     "react-redux": "~5.0.5",
-    "redux": "~3.7.1",
+    "redux": "~3.7.2",
     "rimraf": "~2.6.1",
     "through2": "~2.0.3",
-    "ts-loader": "~2.2.0",
-    "tslint": "~5.4.3",
-    "tslint-react": "~3.0.0",
-    "typescript": "~2.3.4",
-    "webpack": "~3.0.0"
+    "ts-loader": "~2.3.2",
+    "tslint": "~5.5.0",
+    "tslint-react": "~3.1.0",
+    "typescript": "~2.4.2",
+    "webpack": "~3.4.1"
   }
 }
