+++ packages/redux-forms/src/shared/fieldProps.ts
@@ -4,9 +4,10 @@ import {
   compose,
   over,
   lensProp,
+  omit,
 } from 'ramda';
 
-import { Value, Target } from "./getValue";
+import { Value, Target } from './getValue';
 
 
 export type InputProps = {
@@ -26,9 +27,12 @@ export type MetaProps = {
   active: boolean,
 };
 
-export type SeparatedProps = {
-  input: InputProps;
-  meta: MetaProps;
+export type All<T> = T & InputProps & MetaProps;
+
+export type SeparatedProps<T> = {
+  input: InputProps,
+  meta: MetaProps,
+  rest: T,
 };
 
 
@@ -41,6 +45,14 @@ const INPUT_PROPS = [
   'onBlur',
 ];
 
+export type InputProp =
+  | 'checked'
+  | 'name'
+  | 'value'
+  | 'onChange'
+  | 'onFocus'
+  | 'onBlur';
+
 const META_PROPS = [
   'active',
   'dirty',
@@ -49,17 +61,38 @@ const META_PROPS = [
   'visited',
 ];
 
+export type MetaProp =
+  | 'active'
+  | 'dirty'
+  | 'error'
+  | 'touched'
+  | 'visited';
 
-const maybeCheckProps = (all: InputProps): InputProps => {
+const IGNORE_PROPS = [
+  ...INPUT_PROPS,
+  ...META_PROPS,
+  'validate',
+  'normalize',
+  'defaultValue',
+  '_form',
+  '_addField',
+  '_fieldChange',
+  '_fieldFocus',
+  '_fieldBlur',
+];
+
+
+const maybeCheckProps = <T>(all: All<T>): All<T> => {
   if (typeof all.value === 'boolean') {
     return merge(all, { checked: all.value });
   }
   return all;
 };
 
-const separateProps = <T>(all: T & InputProps & MetaProps): SeparatedProps => ({
-  input: pick<InputProps, InputProps>(INPUT_PROPS, all),
-  meta: pick<InputProps, MetaProps>(META_PROPS, all),
+const separateProps = <T>(all: All<T>): SeparatedProps<T> => ({
+  input: pick<All<T>, InputProp>(INPUT_PROPS, all),
+  meta: pick<All<T>, MetaProp>(META_PROPS, all),
+  rest: omit<T>(IGNORE_PROPS, all),
 });
 
 export default compose(separateProps, maybeCheckProps);
